Done all Textgears error handling, tests; organized Textgears tests; 100% coverage
